move duplicateHandlingType() to packages/ingest & re-factor move-granules and sync-granule to use duplicateHandlingType() to determine duplicate handling
